+++ crates/primitives/receipt/src/lib.rs
@@ -1,7 +1,6 @@
 mod from_blockifier;
-mod from_starknet_core;
 mod from_starknet_provider;
-mod to_starknet_core;
+mod into_starknet_core;
 pub use from_blockifier::from_blockifier_execution_info;
 
 use serde::{Deserialize, Serialize};
@@ -18,6 +17,36 @@ pub enum TransactionReceipt {
     Declare(DeclareTransactionReceipt),
     Deploy(DeployTransactionReceipt),
     DeployAccount(DeployAccountTransactionReceipt),
+}
+
+impl From<InvokeTransactionReceipt> for TransactionReceipt {
+    fn from(receipt: InvokeTransactionReceipt) -> Self {
+        TransactionReceipt::Invoke(receipt)
+    }
+}
+
+impl From<L1HandlerTransactionReceipt> for TransactionReceipt {
+    fn from(receipt: L1HandlerTransactionReceipt) -> Self {
+        TransactionReceipt::L1Handler(receipt)
+    }
+}
+
+impl From<DeclareTransactionReceipt> for TransactionReceipt {
+    fn from(receipt: DeclareTransactionReceipt) -> Self {
+        TransactionReceipt::Declare(receipt)
+    }
+}
+
+impl From<DeployTransactionReceipt> for TransactionReceipt {
+    fn from(receipt: DeployTransactionReceipt) -> Self {
+        TransactionReceipt::Deploy(receipt)
+    }
+}
+
+impl From<DeployAccountTransactionReceipt> for TransactionReceipt {
+    fn from(receipt: DeployAccountTransactionReceipt) -> Self {
+        TransactionReceipt::DeployAccount(receipt)
+    }
 }
 
 impl TransactionReceipt {
@@ -323,42 +352,158 @@ mod tests {
     }
 
     #[test]
+    fn test_event_compute_hash_pedersen() {
+        let event = Event {
+            from_address: Felt::from(1),
+            keys: vec![Felt::from(2), Felt::from(3)],
+            data: vec![Felt::from(4), Felt::from(5)],
+        };
+
+        let hash = event.compute_hash_pedersen();
+
+        assert_eq!(hash, Felt::from_hex_unchecked("0x770591674368ef723f40ddef92ee7cf2fcf3e244afa15cd0a703fce83a415c1"));
+    }
+
+    #[test]
+    fn test_event_compute_hash_poseidon() {
+        let event = Event {
+            from_address: Felt::from(1),
+            keys: vec![Felt::from(2), Felt::from(3)],
+            data: vec![Felt::from(4), Felt::from(5)],
+        };
+        let transaction_hash = Felt::from(6);
+
+        let hash = event.compute_hash_poseidon(&transaction_hash);
+
+        assert_eq!(hash, Felt::from_hex_unchecked("0x38be6cfe9175d3f260d1886bf932e1b67415c6c0e8b8a6de565326493c5524f"));
+    }
+
+    #[test]
     fn test_transaction_receipt_compute_hash() {
-        let receipt = TransactionReceipt::Invoke(InvokeTransactionReceipt {
+        let receipt: TransactionReceipt = dummy_invoke_receipt().into();
+        let hash = receipt.compute_hash();
+        assert_eq!(hash, Felt::from_hex_unchecked("0x4ec732a8832cee7ed43a5fb10c20077e8b6d84196ea455ce4b06d27472176e2"));
+
+        let receipt: TransactionReceipt = dummy_l1_handler_receipt().into();
+        let hash = receipt.compute_hash();
+        assert_eq!(hash, Felt::from_hex_unchecked("0x4e34009fa7c50a33edcba2912c8f8195bcebc65d73002b712d0e87e4d9c4425"));
+
+        let receipt: TransactionReceipt = dummy_declare_receipt().into();
+        let hash = receipt.compute_hash();
+        assert_eq!(hash, Felt::from_hex_unchecked("0x4ec732a8832cee7ed43a5fb10c20077e8b6d84196ea455ce4b06d27472176e2"));
+
+        let receipt: TransactionReceipt = dummy_deploy_receipt().into();
+        let hash = receipt.compute_hash();
+        assert_eq!(hash, Felt::from_hex_unchecked("0x4ec732a8832cee7ed43a5fb10c20077e8b6d84196ea455ce4b06d27472176e2"));
+
+        let receipt: TransactionReceipt = dummy_deploy_account_receipt().into();
+        let hash = receipt.compute_hash();
+        assert_eq!(hash, Felt::from_hex_unchecked("0x4ec732a8832cee7ed43a5fb10c20077e8b6d84196ea455ce4b06d27472176e2"));
+    }
+
+    fn dummy_messages() -> Vec<MsgToL1> {
+        vec![
+            MsgToL1 {
+                from_address: Felt::from(1),
+                to_address: Felt::from(2),
+                payload: vec![Felt::from(3), Felt::from(4)],
+            },
+            MsgToL1 {
+                from_address: Felt::from(5),
+                to_address: Felt::from(6),
+                payload: vec![Felt::from(7), Felt::from(8)],
+            },
+        ]
+    }
+
+    fn dummy_events() -> Vec<Event> {
+        vec![
+            Event {
+                from_address: Felt::from(1),
+                keys: vec![Felt::from(2), Felt::from(3)],
+                data: vec![Felt::from(4), Felt::from(5)],
+            },
+            Event {
+                from_address: Felt::from(6),
+                keys: vec![Felt::from(7), Felt::from(8)],
+                data: vec![Felt::from(9), Felt::from(10)],
+            },
+        ]
+    }
+
+    fn dummy_execution_ressources() -> ExecutionResources {
+        ExecutionResources {
+            steps: 1,
+            memory_holes: Some(2),
+            range_check_builtin_applications: Some(3),
+            pedersen_builtin_applications: Some(4),
+            poseidon_builtin_applications: Some(5),
+            ec_op_builtin_applications: Some(6),
+            ecdsa_builtin_applications: Some(7),
+            bitwise_builtin_applications: Some(8),
+            keccak_builtin_applications: Some(9),
+            segment_arena_builtin: Some(10),
+            data_availability: DataAvailabilityResources { l1_gas: 11, l1_data_gas: 12 },
+            // TODO: Change with non-default values when starknet-rs supports it.
+            total_gas_consumed: Default::default(),
+        }
+    }
+
+    pub(crate) fn dummy_invoke_receipt() -> InvokeTransactionReceipt {
+        InvokeTransactionReceipt {
             transaction_hash: Felt::from(1),
             actual_fee: FeePayment { amount: Felt::from(2), unit: PriceUnit::Wei },
-            messages_sent: vec![
-                MsgToL1 {
-                    from_address: Felt::from(3),
-                    to_address: Felt::from(4),
-                    payload: vec![Felt::from(5), Felt::from(6)],
-                },
-                MsgToL1 {
-                    from_address: Felt::from(7),
-                    to_address: Felt::from(8),
-                    payload: vec![Felt::from(9), Felt::from(10)],
-                },
-            ],
-            events: vec![],
-            execution_resources: ExecutionResources {
-                steps: 0,
-                memory_holes: None,
-                range_check_builtin_applications: None,
-                pedersen_builtin_applications: None,
-                poseidon_builtin_applications: None,
-                ec_op_builtin_applications: None,
-                ecdsa_builtin_applications: None,
-                bitwise_builtin_applications: None,
-                keccak_builtin_applications: None,
-                segment_arena_builtin: None,
-                data_availability: DataAvailabilityResources { l1_gas: 11, l1_data_gas: 12 },
-                total_gas_consumed: DataAvailabilityResources { l1_gas: 13, l1_data_gas: 14 },
-            },
+            messages_sent: dummy_messages(),
+            events: dummy_events(),
+            execution_resources: dummy_execution_ressources(),
             execution_result: ExecutionResult::Reverted { reason: "aborted".to_string() },
-        });
+        }
+    }
 
-        let hash = receipt.compute_hash();
+    pub(crate) fn dummy_l1_handler_receipt() -> L1HandlerTransactionReceipt {
+        L1HandlerTransactionReceipt {
+            message_hash: Felt::from(1),
+            transaction_hash: Felt::from(2),
+            actual_fee: FeePayment { amount: Felt::from(3), unit: PriceUnit::Wei },
+            messages_sent: dummy_messages(),
+            events: dummy_events(),
+            execution_resources: dummy_execution_ressources(),
+            execution_result: ExecutionResult::Reverted { reason: "aborted".to_string() },
+        }
+    }
 
-        assert_eq!(hash, Felt::from_hex_unchecked("0x26feec2eb76b63633bc4c1af780813b52a0809a1dae192f53397385da95b8"));
+    pub(crate) fn dummy_declare_receipt() -> DeclareTransactionReceipt {
+        DeclareTransactionReceipt {
+            transaction_hash: Felt::from(1),
+            actual_fee: FeePayment { amount: Felt::from(2), unit: PriceUnit::Wei },
+            messages_sent: dummy_messages(),
+            events: dummy_events(),
+            execution_resources: dummy_execution_ressources(),
+            execution_result: ExecutionResult::Reverted { reason: "aborted".to_string() },
+        }
+    }
+
+    pub(crate) fn dummy_deploy_receipt() -> DeployTransactionReceipt {
+        DeployTransactionReceipt {
+            transaction_hash: Felt::from(1),
+            actual_fee: FeePayment { amount: Felt::from(2), unit: PriceUnit::Wei },
+            messages_sent: dummy_messages(),
+            events: dummy_events(),
+            execution_resources: dummy_execution_ressources(),
+            execution_result: ExecutionResult::Reverted { reason: "aborted".to_string() },
+            contract_address: Felt::from(3),
+        }
+    }
+
+    pub(crate) fn dummy_deploy_account_receipt() -> DeployAccountTransactionReceipt {
+        DeployAccountTransactionReceipt {
+            transaction_hash: Felt::from(1),
+            actual_fee: FeePayment { amount: Felt::from(2), unit: PriceUnit::Wei },
+            messages_sent: dummy_messages(),
+            events: dummy_events(),
+            execution_resources: dummy_execution_ressources(),
+            execution_result: ExecutionResult::Reverted { reason: "aborted".to_string() },
+            contract_address: Felt::from(3),
+        }
     }
 }
