+++ crates/primitives/transactions/src/broadcasted_to_blockifier.rs
@@ -1,4 +1,4 @@
-use crate::{to_starknet_api::TransactionApiError, Transaction, TransactionWithHash};
+use crate::{into_starknet_api::TransactionApiError, Transaction, TransactionWithHash};
 use blockifier::{execution::errors::ContractClassError, transaction::errors::TransactionExecutionError};
 use dp_class::{to_blockifier_class, ClassHash, ClassInfo, ContractClass, ConvertedClass, ToCompiledClass};
 use starknet_api::transaction::TransactionHash;
